move actions to make sense
