+++ CHANGELOG.md
@@ -1,0 +1,6 @@
+# Change log
+
+
+# v0.2.0
+* Moved actions to make sense
+* Add github workflow
