+++ pyawx/__init__.py
@@ -6,4 +6,4 @@ Copyright © 2019, iRunAsRoot
 
 from pyawx.api import Client
 
-__version__ = "0.1.1"
+__version__ = "0.2.0"
