+++ pyawx/models/__init__.py
@@ -1,7 +1,7 @@
 """
-file-name.py - desc
-Comments: 
 Author: Dennis Whitney
 Email: dennis@runasroot.com
 Copyright © 2019, iRunAsRoot 
 """
+
+from ._mixins import DataModelMixin
